refactor(publish): build DocletPage doclet children at instantiation
Instead of iterating through all doclets in the container, assume DocletPage instance has everything needed to render a page.
As such, DocletPage is now instantiated with 'memberOf' children instead of matching 'longname' doclets.
This gives DocletPage everything it needs to render the contents of a full doclet page (e.g. for classes and namespaces, etc).